tg_ngrok_bot_protect/requirements.txt
python-telegram-bot==20.3
pyngrok==6.0.0
